# .context-kit/orchestration/enhanced-mode-bidirectional-highlighting/integration-contracts/openapi-spec.yaml
openapi: 3.0.3
info:
  title: DocuSearch Enhanced Mode API
  description: |
    API endpoints for bidirectional highlighting between research answers and source documents.

    This specification covers new and enhanced endpoints required for the enhanced mode feature:
    - Structure metadata extraction
    - Chunk-level navigation
    - Research API with chunk references

    **Version**: 1.0
    **Last Updated**: 2025-10-17
  version: 1.0.0
  contact:
    name: DocuSearch Development Team
servers:
  - url: http://localhost:8002
    description: Worker API (documents, structure, chunks)
  - url: http://localhost:8003
    description: Research API
paths:
...
  /api/research/ask:
    post:
      summary: Submit research question with AI-generated answer
      description: |
        Enhanced research endpoint that returns AI-generated answers with inline citations
        and chunk-level source references for precise navigation.
      operationId: askResearchQuestion
      tags:
        - Research
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchRequest'
            example:
              query: "What caused the 2008 financial crisis?"
              num_sources: 10
              search_mode: hybrid
      responses:
        '200':
          description: Research answer generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchResponse'
              example:
                answer: "The 2008 financial crisis was caused by several interconnected factors. The primary cause was the collapse of the subprime mortgage market [1]..."
                citations:
                  - id: 1
                    start: 155
                    end: 158
                    marker: "[1]"
                  - id: 2
                    start: 320
                    end: 323
                    marker: "[2]"
                sources:
                  - id: 1
                    doc_id: fin-crisis-2008-report
                    filename: Financial_Crisis_Report_2009.pdf
                    page: 12
                    extension: pdf
                    thumbnail_path: /images/fin-crisis-2008-report/page012_thumb.jpg
                    date_added: "2025-10-15T10:30:00Z"
                    relevance_score: 0.89
                    chunk_id: fin-crisis-2008-report-chunk0045
                  - id: 2
                    doc_id: housing-bubble-analysis
                    filename: Housing_Market_Analysis.pdf
                    page: 8
                    extension: pdf
                    thumbnail_path: /images/housing-bubble-analysis/page008_thumb.jpg
                    date_added: "2025-10-14T15:20:00Z"
                    relevance_score: 0.82
                    chunk_id: null
                citation_map:
                  "1":
                    - sentence_index: 1
                      sentence_text: "The primary cause was the collapse of the subprime mortgage market [1]."
                metadata:
                  total_sources: 5
                  context_tokens: 2850
                  context_truncated: false
                  model_used: gpt-4-turbo
                  search_mode: hybrid
                  processing_time_ms: 2450
                  llm_latency_ms: 1950
                  search_latency_ms: 245
components:
  schemas:
...
    ResearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 3
          maxLength: 500
          description: Research question from user
        num_sources:
          type: integer
          minimum: 1
          maximum: 20
          default: 10
          description: Maximum number of source documents
        search_mode:
          type: string
          enum: [visual, text, hybrid]
          default: hybrid
          description: Search collection mode
        model:
          type: string
          nullable: true
          description: LLM model to use (default from config)
        temperature:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          nullable: true
          description: LLM temperature (default 0.3)
    ResearchResponse:
      type: object
      required:
        - answer
        - citations
        - citation_map
        - sources
        - metadata
      properties:
        answer:
          type: string
          description: AI-generated answer with inline citations
        citations:
          type: array
          items:
            $ref: '#/components/schemas/CitationInfo'
        citation_map:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/SentenceInfo'
          description: Map of citation numbers to sentences
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceInfo'
        metadata:
          $ref: '#/components/schemas/ResearchMetadata'
    CitationInfo:
      type: object
      required:
        - id
        - start
        - end
        - marker
      properties:
        id:
          type: integer
          minimum: 1
          description: Citation number
        start:
          type: integer
          minimum: 0
          description: Start position in answer text
        end:
          type: integer
          minimum: 0
          description: End position in answer text
        marker:
          type: string
          description: Citation marker (e.g., "[1]")
    SentenceInfo:
      type: object
      required:
        - sentence_index
        - sentence_text
      properties:
        sentence_index:
          type: integer
          minimum: 0
          description: Sentence index
        sentence_text:
          type: string
          description: Sentence text with citation
    SourceInfo:
      type: object
      required:
        - id
        - doc_id
        - filename
        - page
        - extension
        - date_added
        - relevance_score
      properties:
        id:
          type: integer
          minimum: 1
          description: Citation number (1-indexed)
        doc_id:
          type: string
          description: Document identifier
        filename:
          type: string
          description: Original filename
        page:
          type: integer
          minimum: 1
          description: Page number
        extension:
          type: string
          description: File extension
        thumbnail_path:
          type: string
          nullable: true
          description: Thumbnail image URL
        date_added:
          type: string
          format: date-time
          description: ISO 8601 timestamp
        relevance_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Search relevance score
        chunk_id:
          type: string
          nullable: true
          description: |
            Chunk identifier for text search results.
            Null for visual search results (page-level only).
            Format: {doc_id}-chunk{NNNN}
    ResearchMetadata:
      type: object
      required:
        - total_sources
        - context_tokens
        - context_truncated
        - model_used
        - search_mode
        - processing_time_ms
        - llm_latency_ms
        - search_latency_ms
      properties:
        total_sources:
          type: integer
          minimum: 0
          description: Number of sources used
        context_tokens:
          type: integer
          minimum: 0
          description: Total tokens in context
        context_truncated:
          type: boolean
          description: Whether context was truncated
        model_used:
          type: string
          description: LLM model name
        search_mode:
          type: string
          description: Search mode used
        processing_time_ms:
          type: integer
          minimum: 0
          description: Total processing time
        llm_latency_ms:
          type: integer
          minimum: 0
          description: LLM inference latency
        search_latency_ms:
          type: integer
          minimum: 0
          description: Search latency
...